refractured common modules into modular_robot._common
