+++ experimentation/revolve2/experimentation/genotypes/cppnwin/modular_robot/v2/_body_develop.py
@@ -4,7 +4,8 @@ from queue import Queue
 from typing import Any
 
 import multineat
-from revolve2.modular_robot import ActiveHinge, Brick, Core, Module
+from revolve2.modular_robot import Module
+from revolve2.modular_robot._common import ActiveHinge, Brick, Core
 from revolve2.modular_robot.v2 import BodyV2
