+++ modular_robot/revolve2/modular_robot/__init__.py
@@ -1,11 +1,9 @@
 """Classes and functions to describe and work with modular robots as used in the CI Group at VU Amsterdam."""
 
-from ._active_hinge import ActiveHinge
+
 from ._body import Body
 from ._body_state import BodyState
 from ._brain import Brain
-from ._brick import Brick
-from ._core import Core
 from ._create_batch_multiple_isolated_robots import (
     create_batch_multiple_isolated_robots,
 )
@@ -21,12 +19,9 @@ from ._not_finalized_error import NotFinalizedError
 from ._right_angles import RightAngles
 
 __all__ = [
-    "ActiveHinge",
     "Body",
     "BodyState",
     "Brain",
-    "Brick",
-    "Core",
     "ModularRobot",
     "Module",
     "MorphologicalMeasures",
