+++ modular_robot/revolve2/modular_robot/_common/__init__.py
@@ -1,0 +1,6 @@
+"""Common Modules for Robots."""
+from ._active_hinge import ActiveHinge
+from ._brick import Brick
+from ._core import Core
+
+__all__ = ["ActiveHinge", "Brick", "Core"]
