+++ modular_robot/revolve2/modular_robot/_morphological_measures.py
@@ -1,9 +1,7 @@
 import numpy as np
 
-from ._active_hinge import ActiveHinge
 from ._body import Body
-from ._brick import Brick
-from ._core import Core
+from ._common import ActiveHinge, Brick, Core
 from ._module import Module
 from ._not_finalized_error import NotFinalizedError
